refactor: re-organize EFB files
