+++ package.json
@@ -19,10 +19,10 @@
     "@babel/core": "^7.12.3",
     "@babel/eslint-parser": "^7.12.1",
     "@babel/plugin-proposal-class-properties": "~7.12.1",
+    "@babel/plugin-transform-runtime": "~7.12.1",
     "@babel/preset-env": "^7.12.1",
     "@babel/preset-react": "^7.12.1",
     "@babel/preset-typescript": "~7.12.7",
-    "@babel/plugin-transform-runtime": "~7.12.1",
     "@rollup/plugin-babel": "^5.2.1",
     "@rollup/plugin-commonjs": "^15.1.0",
     "@rollup/plugin-image": "^2.0.5",
