+++ src/instruments/src/EFB/Assets/Fonts/weather-icons.css
@@ -21,8 +21,8 @@
  */
 @font-face {
   font-family: 'weathericons';
-  src: url('../font/weathericons-regular-webfont.eot');
-  src: url('../font/weathericons-regular-webfont.eot?#iefix') format('embedded-opentype'), url('../font/weathericons-regular-webfont.woff2') format('woff2'), url('../font/weathericons-regular-webfont.woff') format('woff'), url('../font/weathericons-regular-webfont.ttf') format('truetype'), url('../font/weathericons-regular-webfont.svg#weather_iconsregular') format('svg');
+  src: url('src/instruments/src/EFB/Assets/Fonts/weathericons-regular-webfont.eotthericons-regular-webfont.eot');
+  src: url('src/instruments/src/EFB/Assets/Fonts/weathericons-regular-webfont.eotthericons-regular-webfont.eot?#iefix') format('embedded-opentype'), url('src/instruments/src/EFB/Assets/Fonts/weathericons-regular-webfont.woff2ericons-regular-webfont.woff2') format('woff2'), url('src/instruments/src/EFB/Assets/Fonts/weathericons-regular-webfont.woffhericons-regular-webfont.woff') format('woff'), url('src/instruments/src/EFB/Assets/Fonts/weathericons-regular-webfont.ttfthericons-regular-webfont.ttf') format('truetype'), url('src/instruments/src/EFB/Assets/Fonts/weathericons-regular-webfont.svgthericons-regular-webfont.svg#weather_iconsregular') format('svg');
   font-weight: normal;
   font-style: normal;
 }
