+++ src/instruments/src/EFB/Assets/Fonts/weathericons-regular-webfont.ttf
