+++ src/instruments/src/EFB/Dashboard/DashboardWidget.scss
@@ -1,6 +1,3 @@
-@import '../Efb.scss';
-@import './weather-icons.scss';
-
 .DashboardWidget {
     display: grid;
     font-family: "Nunito Sans", sans-serif;
