+++ src/instruments/src/EFB/Dashboard/Widgets/WeatherWidgetIcons.ts
