+++ src/instruments/src/EFB/Loadsheet/LoadsheetWidget.scss
