+++ src/instruments/src/EFB/Loadsheet/LoadsheetWidget.tsx
@@ -1,4 +1,4 @@
-import React, {useState, useEffect} from 'react';
+import React, { useState, useEffect } from 'react';
 
 type LoadsheetWidgetProps = {
     weights: {
