+++ src/instruments/src/EFB/Profile/Profile.scss
@@ -1,5 +1,3 @@
-@import '../Efb.scss';
-
 .ProfileWidgets {
     margin: 20px 30px 30px 30px;
     display: grid;
