+++ src/instruments/src/EFB/Profile/logo192.png
