+++ src/instruments/src/EFB/Settings/Settings.scss
@@ -1,5 +1,3 @@
-@import '../Efb.scss';
-
 .Settings {
     display: flex;
     flex-direction: column;
