+++ src/instruments/src/EFB/ToolBar/ToolBar.scss
@@ -1,5 +1,3 @@
-@import '../Efb.scss';
-
 .Toolbar {
   $padding: 1em;
   $spacing-padding: 2.25rem;
