+++ src/instruments/src/EFB/index.jsx
@@ -16,8 +16,6 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-// eslint-disable-next-line import/no-extraneous-dependencies,no-unused-vars
-import React from 'react';
 import ReactDOM from 'react-dom';
 import Efb from './Efb.tsx';
 import { renderTarget } from '../util.mjs';
